updated according to remarks
