--- src/main/kotlin/AbstractFactoryPattern/Blue.kt
@@ -1,11 +1,0 @@
-package AbstractFactoryPattern
-
-/**
- * Created by Mikhail on 18.09.2015.
- */
-public class Blue : Color {
-
-    override fun fill() {
-        println("Inside Blue::fill() method.")
-    }
-}
