--- src/main/kotlin/AbstractFactoryPattern/Circle.kt
@@ -1,11 +1,0 @@
-package AbstractFactoryPattern
-
-/**
- * Created by Mikhail on 18.09.2015.
- */
-public class Circle : Shape {
-
-    override fun draw() {
-        println("Inside Circle::draw() method.")
-    }
-}
