--- src/main/kotlin/AbstractFactoryPattern/Green.kt
@@ -1,11 +1,0 @@
-package AbstractFactoryPattern
-
-/**
- * Created by Mikhail on 18.09.2015.
- */
-public class Green : Color {
-
-    override fun fill() {
-        println("Inside Green::fill() method.")
-    }
-}
