--- src/main/kotlin/AbstractFactoryPattern/Rectangle.kt
@@ -1,11 +1,0 @@
-package AbstractFactoryPattern
-
-/**
- * Created by Mikhail on 18.09.2015.
- */
-public class Rectangle : Shape {
-
-    override fun draw() {
-        println("Inside Rectangle::draw() method.")
-    }
-}
