--- src/main/kotlin/AbstractFactoryPattern/Red.kt
@@ -1,11 +1,0 @@
-package AbstractFactoryPattern
-
-/**
- * Created by Mikhail on 18.09.2015.
- */
-public class Red : Color {
-
-    override fun fill() {
-        println("Inside Red::fill() method.")
-    }
-}
