--- src/main/kotlin/AbstractFactoryPattern/Shape.kt
@@ -1,8 +1,0 @@
-package AbstractFactoryPattern
-
-/**
- * Created by Mikhail on 18.09.2015.
- */
-public interface Shape {
-    public fun draw()
-}
