--- src/main/kotlin/AbstractFactoryPattern/Square.kt
@@ -1,11 +1,0 @@
-package AbstractFactoryPattern
-
-/**
- * Created by Mikhail on 18.09.2015.
- */
-public class Square : Shape {
-
-    override fun draw() {
-        println("Inside Square::draw() method.")
-    }
-}
